remove storyboard and add folders
