+++ Login Screen/Sources/App/AppDelegate.swift
@@ -1,0 +1,19 @@
+//
+//  AppDelegate.swift
+//  Login Screen
+//
+//  Created by Serhii  on 19/08/2022.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        // Override point for customization after application launch.
+        return true
+    }
+
+}
+
