+++ Login Screen/Sources/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
